imagesArray put in data folder
